+++ tests/Rules/PHPUnit/data/InvalidClassCoversAnnotation.php
@@ -2,6 +2,8 @@
 
 namespace Dummy;
 
+use Rules\Methods\data\Foo;
+
 /**
  * @backupGlobals
  * @backupGlobals enabled
